requirements.txt
requests
openai
pillow
tqdm
opencv-python-headless
numpy
msgpack-rpc-python
airsim